Create Data Science to Analyze Weather Trends with Python and NOAA API
Using Python and NOAA's API, I fetched temperature data at Chicago O'Hare International Airport (GHCND:USW00094846) for the year 2000. With this data in hand, I employed pandas to manipulate and analyze it, and matplotlib to visualize the insights. 



This code is customizable, as a result we can extract any weather stations data for a particular year. The variables are in these case 'Station ID', 'Threshold Temperature' for extreme temperature days and 'Year'



Here's what the script does:

1️⃣ Fetches temperature data from the NOAA API.

2️⃣ Converts the data into a pandas DataFrame.

3️⃣ Calculates the 5-day moving average temperature.

4️⃣ Identifies extreme temperature days based on a specified threshold.

5️⃣ Displays the temperature data in both table and chart formats.



The script visualizes the temperature data in two plots: the first displays the daily temperature variation throughout the year, while the second overlays the 5-day moving average on top of the daily temperature values where red dots are marked as extreme temperature days for that year based on threshold value.